feat!: moved from create-react-app to vite
