+++ package.json
@@ -1,49 +1,35 @@
 {
   "name": "my_diary_v2",
   "version": "0.1.0",
-  "private": true,
+  "private": false,
+  "type": "module",
+  "homepage": "/my_diary",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
+  },
   "dependencies": {
-    "@testing-library/jest-dom": "^5.12.0",
-    "@testing-library/react": "^11.2.6",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/react-transition-group": "^1.1.8",
-    "crypto-browserify": "^3.12.0",
     "dropbox": "^9.7.0",
     "event-js": "^0.1.3",
     "extripo": "^0.8.7",
     "isomorphic-fetch": "^3.0.0",
-    "perlin-simplex": "^0.0.3",
-    "react": "^17.0.2",
-    "react-device-detect": "^1.17.0",
-    "react-dom": "^17.0.2",
-    "react-scripts": "^5.0.1",
+    "react": "^18.2.0",
+    "react-device-detect": "^2.2.3",
+    "react-dom": "^18.2.0",
     "react-tag-input-component": "^2.0.2",
     "react-transition-group": "^4.4.5",
-    "util": "^0.12.5",
-    "web-vitals": "^1.1.2"
+    "web-vitals": "^3.5.0"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "devDependencies": {
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
+    "@vitejs/plugin-react": "^4.0.3",
+    "eslint": "^8.45.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
+    "vite": "^4.4.5"
   }
 }
