--- src/components/elements.js
@@ -1,14 +1,0 @@
-import React from 'react'
-
-export class Label extends React.Component{
-    constructor(props){
-        super(props)
-    }
-    render(){
-        return (
-            <div className="label">
-                {this.props.text}
-            </div>
-        )
-    }
-}
