+++ vite.config.js
@@ -1,0 +1,6 @@
+import { defineConfig } from 'vite'
+import packageInfo from './package.json'
+
+export default defineConfig({
+    base: packageInfo.homepage,
+})
